refactor utils

+++ x-pack/plugins/osquery/server/routes/saved_query/delete_saved_query_route.ts
@@ -7,10 +7,10 @@
 
 import { schema } from '@kbn/config-schema';
 import { IRouter } from '@kbn/core/server';
-import { find } from 'lodash';
-import { OSQUERY_INTEGRATION_NAME, PLUGIN_ID } from '../../../common';
+import { PLUGIN_ID } from '../../../common';
 import { savedQuerySavedObjectType } from '../../../common/types';
 import { OsqueryAppContext } from '../../lib/osquery_app_context_services';
+import { isSavedQueryPrebuilt } from './utils';
 
 export const deleteSavedQueryRoute = (router: IRouter, osqueryContext: OsqueryAppContext) => {
   router.delete(
@@ -27,18 +27,9 @@ export const deleteSavedQueryRoute = (router: IRouter, osqueryContext: OsqueryAp
       const coreContext = await context.core;
       const savedObjectsClient = coreContext.savedObjects.client;
 
-      const installation = await osqueryContext.service
-        .getPackageService()
-        ?.asInternalUser?.getInstallation(OSQUERY_INTEGRATION_NAME);
-
-      if (installation) {
-        const installationSavedQueries = find(
-          installation.installed_kibana,
-          (item) => item.type === savedQuerySavedObjectType && item.id === request.params.id
-        );
-        if (installationSavedQueries) {
-          return response.conflict({ body: `Elastic prebuilt Saved query cannot be deleted.` });
-        }
+      const isPrebuilt = await isSavedQueryPrebuilt(osqueryContext, request.params.id);
+      if (isPrebuilt) {
+        return response.conflict({ body: `Elastic prebuilt Saved query cannot be deleted.` });
       }
 
       await savedObjectsClient.delete(savedQuerySavedObjectType, request.params.id, {
